Refactor player choice function and game function
The player can now simply click a button for selection
instead of typing in a prompt as repeatedly typing
can get boring. Also, the game function now displays
the result of each round on the webpage instead of
on the console so they don't have to open the
browser developer tools to see the results of each
round.